added camera controller

+++ src/input.h
@@ -30,10 +30,10 @@ public:
     Event<> quit;
     Event<SDL_Keycode, uint16_t> keyDown;
     Event<SDL_Keycode, uint16_t> keyUp;
-    Event<uint8_t, uint8_t, int32_t, int32_t> mouseButtonDown;
-    Event<uint8_t, int32_t, int32_t> mouseButtonUp;
-    Event<int32_t, int32_t> mouseMotion;
-    Event<int32_t> mouseWheel;
+    Event<uint8_t, uint8_t, float, float> mouseButtonDown;
+    Event<uint8_t, float, float> mouseButtonUp;
+    Event<float, float> mouseMotion;
+    Event<float> mouseWheel;
 };
 }
